tweaks

+++ taichi/backends/device.h
@@ -438,7 +438,7 @@ class Device {
   void print_all_cap() const;
 
   struct AllocParams {
-    uint64_t size{0};
+    size_t size{0};
     bool host_write{false};
     bool host_read{false};
     bool export_sharing{false};
